requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorlog==6.7.0
frozenlist==1.3.3
h11==0.14.0
hikari==2.0.0.dev113
hikari-lightbulb==2.3.1
idna==3.4
multidict==6.0.3
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
yarl==1.8.2
